WIP animation: finish animating tip
I faced two struggles to get this working:

a) The tip itself is in a clipping group. Some of the animation applies
to the group and some applies to the tip itself. Especially the rotation
point is in relation to the surrounding group.

b) The order of the container groups in the robot needed to change, so
that the tip is visible during liquid handling.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>